requirements.txt
Django==3.0.7
# testing
pytest-django==4.1.0
